--- .eslintrc.js
@@ -1,12 +1,0 @@
-module.exports = {
-  plugins: ['prettier'],
-  env: {
-    browser: true,
-    es2021: true,
-  },
-  extends: ['airbnb-base'],
-  parserOptions: {
-    ecmaVersion: 13,
-    sourceType: 'module',
-  },
-};
