--- .gitignore
@@ -1,7 +1,0 @@
-santasprites/
-
-node_modules/
-
-dist/
-
-.cache/
