--- .prettierrc
@@ -1,7 +1,0 @@
-{
-  "tabWidth": 2,
-  "useTabs": false,
-  "singleQuote": true,
-  "trailingComma": "all",
-  "printWidth": 100
-}
