+++ README.md
@@ -1,2 +1,37 @@
-# rudolph-run
-크리스마스를 축하하는 게임입니다!
+# 🎅루돌프 런
+
+크리스마스를 기념하여 만든 정~말 간단한 게임입니다. Merry Christmas~🎄
+
+## ✨프로젝트 목표
+
+루돌프 런 프로젝트를 통해 아래와 같은 기술들을 학습하였습니다.
+
+- Canvas API
+- 프로젝트 셋팅
+  - Prettier
+  - Eslint
+  - Parcel(Bundler)
+- 배포 및 도메인 설정(<a>rudolph-run.ga</a>)
+  - Amazon S3
+  - Amazon CloudFront
+  - Amazon Route53
+  - Amazon Certificate Manager
+
+## 🎮인 게임
+
+### 좌우 이동
+
+![move](./image/move.gif)
+
+### 점프
+![jump](./image/jump.gif)
+
+### 이벤트 컷신
+![event](./image/event.gif)
+
+## 💾설치 및 실행 방법
+로컬 환경에서 실행하는 경우 localhost:1234로 접속할 수 있습니다.
+```
+npm install
+npm start
+```
