+++ app/.eslintrc.js
@@ -1,0 +1,12 @@
+module.exports = {
+  plugins: ['prettier'],
+  env: {
+    browser: true,
+    es2021: true,
+  },
+  extends: ['airbnb-base'],
+  parserOptions: {
+    ecmaVersion: 13,
+    sourceType: 'module',
+  },
+};
