+++ app/.gitignore
@@ -1,0 +1,7 @@
+santasprites/
+
+node_modules/
+
+dist/
+
+.cache/
