+++ app/src/main.js
@@ -1,0 +1,4 @@
+import Game from './Game';
+
+const game = new Game();
+game.init();
