--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "rudolph-run",
-  "version": "1.0.0",
-  "description": "크리스마스를 축하하는 게임입니다!",
-  "main": "main.js",
-  "scripts": {
-    "dev": "parcel src/index.html"
-  },
-  "devDependencies": {
-    "eslint": "^8.5.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.3",
-    "parcel-bundler": "^1.12.5"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jaesoekjjang/rudolph-run.git"
-  },
-  "author": "jaesoekjjang",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/jaesoekjjang/rudolph-run/issues"
-  },
-  "homepage": "https://github.com/jaesoekjjang/rudolph-run#readme"
-}
