--- src/index.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Santa Rider</title>
-    <style>
-      body {
-        padding: 0;
-        margin: 0;
-      }
-
-      #canvas {
-        width: 1600px;
-        heieght: 1000px;
-        border: 2px solid black;
-        margin: 0;
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%) scale(0.5);
-      }
-
-      img {
-        display: none;
-      }
-    </style>
-  </head>
-  <body>
-    <canvas id="canvas" width="800" height="500"></canvas>
-    <script src="main.js" async></script>
-  </body>
-</html>
