--- src/main.js
@@ -1,4 +1,0 @@
-import Game from './Game';
-
-const game = new Game();
-game.init();
